fix some more type conversion warnings w/MSVC
API & memory related warnings will be dealt with separately

+++ src/dec/idec.c
@@ -98,8 +98,8 @@ static int AppendToMemBuffer(WebPIDecoder* const idec,
   if (mem->end_ + data_size > mem->buf_size_) {  // Need some free memory
     int p;
     uint8_t* new_buf = NULL;
-    const int num_chunks = (MemDataSize(mem) + data_size + CHUNK_SIZE - 1)
-        / CHUNK_SIZE;
+    const size_t num_chunks =
+        (MemDataSize(mem) + data_size + CHUNK_SIZE - 1) / CHUNK_SIZE;
     const size_t new_size = num_chunks * CHUNK_SIZE;
     const uint8_t* const base = mem->buf_ + mem->start_;
